requirements.txt
# For the Hedgehog 2D application
pygame
# For tile image creation script
pillow
